+++ palace/fem/coefficient.hpp
@@ -30,7 +30,7 @@ namespace palace
 
 // Base class for coefficients which need to evaluate a GridFunction in a domain element
 // attached to a boundary element, or both domain elements on either side for internal
-// boundaries.
+// boundaries, with optional scaling factor.
 class BdrGridFunctionCoefficient
 {
 protected:
@@ -39,6 +39,7 @@ protected:
   const mfem::ParMesh &mesh;
   mfem::FaceElementTransformations FET;
   mfem::IsoparametricTransformation T1, T2;
+  const double scaling;  // scaling factor used for unit conversions
 
   bool GetBdrElementNeighborTransformations(int i, const mfem::IntegrationPoint &ip)
   {
@@ -48,7 +49,10 @@ protected:
   }
 
 public:
-  BdrGridFunctionCoefficient(const mfem::ParMesh &mesh) : mesh(mesh) {}
+  BdrGridFunctionCoefficient(const mfem::ParMesh &mesh, double scaling = 1.0)
+    : mesh(mesh), scaling(scaling)
+  {
+  }
 
   // For a boundary element, return the element transformation objects for the neighboring
   // domain elements. FET.Elem2 may be nullptr if the boundary is a true one-sided boundary,
@@ -87,9 +91,10 @@ private:
 
 public:
   BdrSurfaceCurrentVectorCoefficient(const mfem::ParGridFunction &B,
-                                     const MaterialOperator &mat_op)
+                                     const MaterialOperator &mat_op, double scaling = 1.0)
     : mfem::VectorCoefficient(B.VectorDim()),
-      BdrGridFunctionCoefficient(*B.ParFESpace()->GetParMesh()), B(B), mat_op(mat_op)
+      BdrGridFunctionCoefficient(*B.ParFESpace()->GetParMesh(), scaling), B(B),
+      mat_op(mat_op)
   {
   }
 
@@ -124,10 +129,12 @@ public:
     GetNormal(T, normal, ori);
     V.SetSize(vdim);
     linalg::Cross3(normal, VU, V);
+
+    V *= scaling;
   }
 };
 
-// Computes the flux Φₛ = F ⋅ n with F = B or ε D on interior boundary elements using B or
+// Computes the flux Φₛ = F ⋅ n with F = B or ε E on interior boundary elements using B or
 // E given as a vector grid function. For a two-sided internal boundary, the contributions
 // from both sides can either add or be averaged.
 template <SurfaceFlux Type>
@@ -139,15 +146,15 @@ private:
   const MaterialOperator &mat_op;
   bool two_sided;
   const mfem::Vector &x0;
-
   void GetLocalFlux(mfem::ElementTransformation &T, mfem::Vector &V) const;
 
 public:
   BdrSurfaceFluxCoefficient(const mfem::ParGridFunction *E, const mfem::ParGridFunction *B,
                             const MaterialOperator &mat_op, bool two_sided,
-                            const mfem::Vector &x0)
-    : mfem::Coefficient(), BdrGridFunctionCoefficient(E ? *E->ParFESpace()->GetParMesh()
-                                                        : *B->ParFESpace()->GetParMesh()),
+                            const mfem::Vector &x0, double scaling = 1.0)
+    : mfem::Coefficient(),
+      BdrGridFunctionCoefficient(
+          E ? *E->ParFESpace()->GetParMesh() : *B->ParFESpace()->GetParMesh(), scaling),
       E(E), B(B), mat_op(mat_op), two_sided(two_sided), x0(x0)
   {
     MFEM_VERIFY((E || (Type != SurfaceFlux::ELECTRIC && Type != SurfaceFlux::POWER)) &&
@@ -219,6 +226,7 @@ inline void BdrSurfaceFluxCoefficient<SurfaceFlux::ELECTRIC>::GetLocalFlux(
   mfem::Vector W(W_data, T.GetSpaceDim());
   E->GetVectorValue(T, T.GetIntPoint(), W);
   mat_op.GetPermittivityReal(T.Attribute).Mult(W, V);
+  V *= scaling;
 }
 
 template <>
@@ -227,6 +235,7 @@ inline void BdrSurfaceFluxCoefficient<SurfaceFlux::MAGNETIC>::GetLocalFlux(
 {
   // Flux B.
   B->GetVectorValue(T, T.GetIntPoint(), V);
+  V *= scaling;
 }
 
 template <>
@@ -242,6 +251,7 @@ BdrSurfaceFluxCoefficient<SurfaceFlux::POWER>::GetLocalFlux(mfem::ElementTransfo
   E->GetVectorValue(T, T.GetIntPoint(), W1);
   V.SetSize(W1.Size());
   linalg::Cross3(W1, W2, V);
+  V *= scaling;
 }
 
 // Computes a single-valued α Eᵀ E on boundaries from E given as a vector grid function.
@@ -452,12 +462,13 @@ class EnergyDensityCoefficient : public mfem::Coefficient, public BdrGridFunctio
 private:
   const GridFunction &U;
   const MaterialOperator &mat_op;
-
   double GetLocalEnergyDensity(mfem::ElementTransformation &T) const;
 
 public:
-  EnergyDensityCoefficient(const GridFunction &U, const MaterialOperator &mat_op)
-    : mfem::Coefficient(), BdrGridFunctionCoefficient(*U.ParFESpace()->GetParMesh()), U(U),
+  EnergyDensityCoefficient(const GridFunction &U, const MaterialOperator &mat_op,
+                           double scaling = 1.0)
+    : mfem::Coefficient(),
+      BdrGridFunctionCoefficient(*U.ParFESpace()->GetParMesh(), scaling), U(U),
       mat_op(mat_op)
   {
   }
@@ -504,7 +515,7 @@ inline double EnergyDensityCoefficient<EnergyDensityType::ELECTRIC>::GetLocalEne
     U.Imag().GetVectorValue(T, T.GetIntPoint(), V);
     dot += mat_op.GetPermittivityReal(T.Attribute).InnerProduct(V, V);
   }
-  return 0.5 * dot;
+  return 0.5 * dot * scaling;
 }
 
 template <>
@@ -520,7 +531,7 @@ inline double EnergyDensityCoefficient<EnergyDensityType::MAGNETIC>::GetLocalEne
     U.Imag().GetVectorValue(T, T.GetIntPoint(), V);
     dot += mat_op.GetInvPermeability(T.Attribute).InnerProduct(V, V);
   }
-  return 0.5 * dot;
+  return 0.5 * dot * scaling;
 }
 
 // Compute time-averaged Poynting vector Re{E x H⋆}, without the typical factor of 1/2. For
@@ -548,13 +559,15 @@ private:
       E.Imag().GetVectorValue(T, T.GetIntPoint(), W1);
       linalg::Cross3(W1, W2, V, true);
     }
+    V *= scaling;
   }
 
 public:
   PoyntingVectorCoefficient(const GridFunction &E, const GridFunction &B,
-                            const MaterialOperator &mat_op)
+                            const MaterialOperator &mat_op, double scaling = 1.0)
     : mfem::VectorCoefficient(E.VectorDim()),
-      BdrGridFunctionCoefficient(*E.ParFESpace()->GetParMesh()), E(E), B(B), mat_op(mat_op)
+      BdrGridFunctionCoefficient(*E.ParFESpace()->GetParMesh(), scaling), E(E), B(B),
+      mat_op(mat_op)
   {
   }