+++ palace/models/postoperator.cpp
@@ -96,11 +96,13 @@ PostOperator<solver_t>::PostOperator(const IoData &iodata, fem_op_t<solver_t> &f
   // Add wave port boundary mode postprocessing, if available.
   if constexpr (std::is_same_v<fem_op_t<solver_t>, SpaceOperator>)
   {
+    // Add scaling factor to output dimensional wave port electric fields.
+    const double scaling = units.Dimensionalize<Units::ValueType::FIELD_E>(1.0);
     for (const auto &[idx, data] : fem_op->GetWavePortOp())
     {
       auto ret = port_E0.emplace(idx, WavePortFieldData());
-      ret.first->second.E0r = data.GetModeFieldCoefficientReal();
-      ret.first->second.E0i = data.GetModeFieldCoefficientImag();
+      ret.first->second.E0r = data.GetModeFieldCoefficientReal(scaling);
+      ret.first->second.E0i = data.GetModeFieldCoefficientImag(scaling);
     }
   }
 
@@ -176,46 +178,76 @@ void PostOperator<solver_t>::SetupFieldCoefficients()
 
   if constexpr (HasEGridFunction<solver_t>())
   {
+    // If E is dimensionalized when the coefficients are evaluated, the scaling only needs
+    // to account for the remaining ε_0 = D / E. This assumes ProjectCoefficient(),
+    // ProjectBdrCoefficient(), or paraview->Save() for U_e, Q_sr, and Q_si are always
+    // called after the E GridFunction has been dimensionalized. To output nondimensional
+    // coefficients, omit the scaling argument and make sure E is nondimensional when the
+    // coefficients are evaluated.
+    const double scaling = units.Dimensionalize<Units::ValueType::FIELD_D>(1.0) /
+                           units.Dimensionalize<Units::ValueType::FIELD_E>(1.0);
+
     // Electric Energy Density.
+    // U_e = 1/2 Dᴴ E = 1/2 ε_0 Eᴴ E.
     U_e = std::make_unique<EnergyDensityCoefficient<EnergyDensityType::ELECTRIC>>(
-        *E, fem_op->GetMaterialOp());
+        *E, fem_op->GetMaterialOp(), scaling);
 
     // Electric Boundary Field & Surface Charge.
     E_sr = std::make_unique<BdrFieldVectorCoefficient>(E->Real());
+    // Q_s = D ⋅ n = ε_0 E ⋅ n.
     Q_sr = std::make_unique<BdrSurfaceFluxCoefficient<SurfaceFlux::ELECTRIC>>(
-        &E->Real(), nullptr, fem_op->GetMaterialOp(), true, mfem::Vector());
+        &E->Real(), nullptr, fem_op->GetMaterialOp(), true, mfem::Vector(), scaling);
 
     if constexpr (HasComplexGridFunction<solver_t>())
     {
       E_si = std::make_unique<BdrFieldVectorCoefficient>(E->Imag());
       Q_si = std::make_unique<BdrSurfaceFluxCoefficient<SurfaceFlux::ELECTRIC>>(
-          &E->Imag(), nullptr, fem_op->GetMaterialOp(), true, mfem::Vector());
+          &E->Imag(), nullptr, fem_op->GetMaterialOp(), true, mfem::Vector(), scaling);
     }
   }
 
   if constexpr (HasBGridFunction<solver_t>())
   {
+    // If B is dimensionalized when the coefficients are evaluated, the scaling only needs
+    // to account for the remaining μ⁻¹ = H / B. This assumes ProjectCoefficient(),
+    // ProjectBdrCoefficient(), or paraview->Save() for U_m, J_sr, and J_si are always
+    // called after the B GridFunction has been dimensionalized. To output nondimensional
+    // coefficients, omit the scaling argument and make sure B is nondimensional when the
+    // coefficients are evaluated.
+    const double scaling = units.Dimensionalize<Units::ValueType::FIELD_H>(1.0) /
+                           units.Dimensionalize<Units::ValueType::FIELD_B>(1.0);
+
     // Magnetic Energy Density.
+    // U_m = 1/2 Hᴴ B = 1/2 μ⁻¹ Bᴴ B.
     U_m = std::make_unique<EnergyDensityCoefficient<EnergyDensityType::MAGNETIC>>(
-        *B, fem_op->GetMaterialOp());
+        *B, fem_op->GetMaterialOp(), scaling);
 
     // Magnetic Boundary Field & Surface Current.
     B_sr = std::make_unique<BdrFieldVectorCoefficient>(B->Real());
-    J_sr = std::make_unique<BdrSurfaceCurrentVectorCoefficient>(B->Real(),
-                                                                fem_op->GetMaterialOp());
+    // J_s = n x H = n x μ⁻¹ B.
+    J_sr = std::make_unique<BdrSurfaceCurrentVectorCoefficient>(
+        B->Real(), fem_op->GetMaterialOp(), scaling);
 
     if constexpr (HasComplexGridFunction<solver_t>())
     {
       B_si = std::make_unique<BdrFieldVectorCoefficient>(B->Imag());
-      J_si = std::make_unique<BdrSurfaceCurrentVectorCoefficient>(B->Imag(),
-                                                                  fem_op->GetMaterialOp());
+      J_si = std::make_unique<BdrSurfaceCurrentVectorCoefficient>(
+          B->Imag(), fem_op->GetMaterialOp(), scaling);
     }
   }
 
   if constexpr (HasEGridFunction<solver_t>() && HasBGridFunction<solver_t>())
   {
     // Poynting Vector.
-    S = std::make_unique<PoyntingVectorCoefficient>(*E, *B, fem_op->GetMaterialOp());
+    // S = Re{E x H⋆} = Re{E x μ⁻¹B⋆}.
+    // E and B will be dimensionalized when the coefficient is evaluated, so the scaling
+    // only needs to account for the remaining μ⁻¹ = H / B. As mentioned above,
+    // ProjectCoefficient(*S.get()) or paraview->Save() should always be called after
+    // E and B have been dimensionalized.
+    const double scaling = units.Dimensionalize<Units::ValueType::FIELD_H>(1.0) /
+                           units.Dimensionalize<Units::ValueType::FIELD_B>(1.0);
+    S = std::make_unique<PoyntingVectorCoefficient>(*E, *B, fem_op->GetMaterialOp(),
+                                                    scaling);
   }
 }
 
@@ -364,11 +396,9 @@ void PostOperator<solver_t>::InitializeParaviewDataCollection(
   }
 }
 
-namespace
-{
-
-template <typename T>
-void ScaleGridFunctions(double L, int dim, bool imag, T &E, T &B, T &V, T &A)
+void ScaleGridFunctions(double L, int dim, std::unique_ptr<GridFunction> &E,
+                        std::unique_ptr<GridFunction> &B, std::unique_ptr<GridFunction> &V,
+                        std::unique_ptr<GridFunction> &A)
 {
   // For fields on H(curl) and H(div) spaces, we "undo" the effect of redimensionalizing
   // the mesh which would carry into the fields during the mapping from reference to
@@ -383,7 +413,7 @@ void ScaleGridFunctions(double L, int dim, bool imag, T &E, T &B, T &V, T &A)
     // Piola transform: J^-T
     E->Real() *= L;
     E->Real().FaceNbrData() *= L;
-    if (imag)
+    if (E->HasImag())
     {
       E->Imag() *= L;
       E->Imag().FaceNbrData() *= L;
@@ -395,7 +425,7 @@ void ScaleGridFunctions(double L, int dim, bool imag, T &E, T &B, T &V, T &A)
     const auto Ld = std::pow(L, dim - 1);
     B->Real() *= Ld;
     B->Real().FaceNbrData() *= Ld;
-    if (imag)
+    if (B->HasImag())
     {
       B->Imag() *= Ld;
       B->Imag().FaceNbrData() *= Ld;
@@ -409,7 +439,51 @@ void ScaleGridFunctions(double L, int dim, bool imag, T &E, T &B, T &V, T &A)
   }
 }
 
-}  // namespace
+void DimensionalizeGridFunctions(Units &units, std::unique_ptr<GridFunction> &E,
+                                 std::unique_ptr<GridFunction> &B,
+                                 std::unique_ptr<GridFunction> &V,
+                                 std::unique_ptr<GridFunction> &A)
+{
+  if (E)
+  {
+    units.DimensionalizeInPlace<Units::ValueType::FIELD_E>(*E);
+  }
+  if (B)
+  {
+    units.DimensionalizeInPlace<Units::ValueType::FIELD_B>(*B);
+  }
+  if (A)
+  {
+    units.DimensionalizeInPlace<Units::ValueType::CURRENT>(A->Real());
+  }
+  if (V)
+  {
+    units.DimensionalizeInPlace<Units::ValueType::VOLTAGE>(V->Real());
+  }
+}
+
+void NondimensionalizeGridFunctions(Units &units, std::unique_ptr<GridFunction> &E,
+                                    std::unique_ptr<GridFunction> &B,
+                                    std::unique_ptr<GridFunction> &V,
+                                    std::unique_ptr<GridFunction> &A)
+{
+  if (E)
+  {
+    units.NondimensionalizeInPlace<Units::ValueType::FIELD_E>(*E);
+  }
+  if (B)
+  {
+    units.NondimensionalizeInPlace<Units::ValueType::FIELD_B>(*B);
+  }
+  if (A)
+  {
+    units.NondimensionalizeInPlace<Units::ValueType::CURRENT>(A->Real());
+  }
+  if (V)
+  {
+    units.NondimensionalizeInPlace<Units::ValueType::VOLTAGE>(V->Real());
+  }
+}
 
 template <ProblemType solver_t>
 void PostOperator<solver_t>::WriteParaviewFields(double time, int step)
@@ -422,8 +496,8 @@ void PostOperator<solver_t>::WriteParaviewFields(double time, int step)
   // visualization. Write the mesh coordinates in the same units as originally input.
   mfem::ParMesh &mesh = E ? *E->ParFESpace()->GetParMesh() : *B->ParFESpace()->GetParMesh();
   mesh::DimensionalizeMesh(mesh, mesh_Lc0);
-  ScaleGridFunctions(mesh_Lc0, mesh.Dimension(), HasComplexGridFunction<solver_t>(), E, B,
-                     V, A);
+  ScaleGridFunctions(mesh_Lc0, mesh.Dimension(), E, B, V, A);
+  DimensionalizeGridFunctions(units, E, B, V, A);
   paraview->SetCycle(step);
   paraview->SetTime(time);
   paraview_bdr->SetCycle(step);
@@ -431,8 +505,8 @@ void PostOperator<solver_t>::WriteParaviewFields(double time, int step)
   paraview->Save();
   paraview_bdr->Save();
   mesh::NondimensionalizeMesh(mesh, mesh_Lc0);
-  ScaleGridFunctions(1.0 / mesh_Lc0, mesh.Dimension(), HasComplexGridFunction<solver_t>(),
-                     E, B, V, A);
+  ScaleGridFunctions(1.0 / mesh_Lc0, mesh.Dimension(), E, B, V, A);
+  NondimensionalizeGridFunctions(units, E, B, V, A);
   Mpi::Barrier(fem_op->GetComm());
 }
 
@@ -533,9 +607,8 @@ void PostOperator<solver_t>::WriteMFEMGridFunctions(double time, int step)
   // visualization. Write the mesh coordinates in the same units as originally input.
   mfem::ParMesh &mesh = E ? *E->ParFESpace()->GetParMesh() : *B->ParFESpace()->GetParMesh();
   mesh::DimensionalizeMesh(mesh, mesh_Lc0);
-  ScaleGridFunctions(mesh_Lc0, mesh.Dimension(), HasComplexGridFunction<solver_t>(), E, B,
-                     V, A);
-
+  ScaleGridFunctions(mesh_Lc0, mesh.Dimension(), E, B, V, A);
+  DimensionalizeGridFunctions(units, E, B, V, A);
   // Create grid function for vector coefficients.
   mfem::ParFiniteElementSpace &fespace = E ? *E->ParFESpace() : *B->ParFESpace();
   mfem::ParGridFunction gridfunc_vector(&fespace);
@@ -632,8 +705,8 @@ void PostOperator<solver_t>::WriteMFEMGridFunctions(double time, int step)
   }
 
   mesh::NondimensionalizeMesh(mesh, mesh_Lc0);
-  ScaleGridFunctions(1.0 / mesh_Lc0, mesh.Dimension(), HasComplexGridFunction<solver_t>(),
-                     E, B, V, A);
+  ScaleGridFunctions(1.0 / mesh_Lc0, mesh.Dimension(), E, B, V, A);
+  NondimensionalizeGridFunctions(units, E, B, V, A);
   Mpi::Barrier(fem_op->GetComm());
 }